add validation and logger

+++ api/controllers/players.js
@@ -1,42 +1,60 @@
 /* eslint-disable new-cap */
 const Players = require('../Models/Players')
+const { validationResult } = require('express-validator')
 const infoMessages = require('../constants/infoMessages')
+const { playersLogger } = require('../logger/logger')
 
-async function create(req, res, next) {
+const create = async(req, res, next) => {
   try {
+    const errors = validationResult(req)
+    console.log(errors)
+    if (!errors.isEmpty()) {
+      console.log('if')
+      playersLogger.error(errors)
+      res.status(403)
+        .json({ errors: errors.array() })
+      return
+    }
     const { name, email, password, age, gamesPlayed, gamesWin, gamesLost } = req.body
     const newPlayer = Players({ name, email, password, age, gamesPlayed, gamesWin, gamesLost })
     await newPlayer.save()
-    console.log(infoMessages.NEW_PLAYER)
+    playersLogger.info(infoMessages.NEW_PLAYER)
     res.status(201)
       .send(newPlayer)
   } catch (error) {
-    console.log(error(error.message, { metadata: error.stack }))
+    playersLogger.error((error.message, { metadata: error.stack }))
     return next(error)
   }
 }
 
-async function fetch(req, res) {
+const fetch = async(req, res) => {
   try {
     const data = await Players.find({})
-    console.log(infoMessages.GET_PLAYERS)
+    playersLogger.info(infoMessages.GET_PLAYERS)
     res.json(data)
   } catch (error) {
-    console.log((error.message, { metadata: error.stack }))
+    playersLogger.error((error.message, { metadata: error.stack }))
     res.status(500)
       .send(error.message)
   }
 }
 
-async function deletePlayer(req, res) {
+const deletePlayer = async(req, res) => {
   try {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      playersLogger.error(errors)
+      res.status(403)
+        .json({ errors: errors.array() })
+      return
+    }
     const playerId = req.params.id
     await Players.findByIdAndDelete(playerId)
-    console.log(infoMessages.DELETE_PLAYER)
+    playersLogger.info(infoMessages.DELETE_PLAYER)
     res.status(204)
       .send()
   } catch (error) {
-    console.log(error.message, { metadata: error.stack })
+    playersLogger.error(error.message, { metadata: error.stack })
     res.status(500)
       .send(error.message)
   }
